add mit license and original creator attribution to test_scale.py: Copyright (c) 2021 Daniel Robertson
